fix: moved jwt functions to authMiddleware and jwtMiddleware files. Fixed imports in server.js
